Create a PasswordResetMigrationGenerator
The generator is now invoked by the install generator by default, but can also
be used as a standalone thing.

Also abstracted migration-related methods out from InstallGenerator so that we
could reuse the `copy_migrations` method. This led to remove several private
methods that seem to be redundant to what Rails generators already give us.